[WRFE-15](refactor): List에 있는 Form 컴포넌트화
